Auto Code Cleanup + Gear Failsafe
The autonomous has been condensed by merging all of the shared code, and just using to case-switch to determine whether some steps are skipped or not. Also, a system has been put in place to (at the start of teleop) close the doors if the sensor is not being tripped.